LongitudeDeg перевели в decimal

+++ Airports.DAL/Entityes/AirportDBModel.cs
@@ -40,7 +40,7 @@ namespace Airports.DAL.Entityes
         /// <summary>
         /// Долгота аэропорта в градусах (положительная для востока).
         /// </summary>
-        public double? LongitudeDeg { get; set; }
+        public decimal? LongitudeDeg { get; set; }
 
         /// <summary>
         /// Высота аэропорта над уровнем моря в футах ( не в метрах).
